OB-34: Rename auth-token.js to auth.js
