+++ app/reducers/index.js
@@ -2,7 +2,7 @@ import { combineReducers } from 'redux';
 
 import appData from './app-data';
 import editEntry from './entry';
-import authToken from './auth-token';
+import authToken from './auth';
 
 const appReducers = combineReducers({
     appData,
